# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/properties/draft": {
      "post": {
        "operationId": "PropertiesController_createDraft",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new property onboarding draft",
        "tags": [
          "Properties"
        ]
      }
    },
    "/properties/{propertyId}/step/{stepNum}": {
      "patch": {
        "operationId": "PropertiesController_saveStep",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stepNum",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Auto-save a wizard step (1-5) for a draft property",
        "tags": [
          "Properties"
        ]
      }
    },
    "/properties/{propertyId}/draft": {
      "get": {
        "operationId": "PropertiesController_getDraft",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the current onboarding draft state",
        "tags": [
          "Properties"
        ]
      }
    },
    "/properties/{propertyId}/submit": {
      "post": {
        "operationId": "PropertiesController_submit",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit a draft property for review",
        "tags": [
          "Properties"
        ]
      }
    },
    "/properties/{propertyId}/status": {
      "get": {
        "operationId": "PropertiesController_getStatus",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the submission status and review timeline",
        "tags": [
          "Properties"
        ]
      }
    },
    "/properties/{propertyId}/revise": {
      "patch": {
        "operationId": "PropertiesController_revise",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resubmit a property after a requested revision",
        "tags": [
          "Properties"
        ]
      }
    },
...
  },
  "info": {
    "title": "QuickNest API",
    "description": "Backend APIs for QuickNest hotel booking platform",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}